requirements.txt
flask==2.0.1
flask-socketio==5.1.1
eventlet==0.33.0
pyngrok==5.1.0
python-dotenv==0.19.1
flask-cors==3.0.10
